s3: Added support for S3 Bucket as installation source
The installation in AWS could be done with Oracle Images placed on S3.
The following variables will mount the S3 Bucket with s2fs-fuse directly to the OS.
Do not forget to add a costly free endpoint for S3 to your VPC. Otherwise the traffic could
run over the the expensive NAT Gateway.

Parameter:

  s3_bucket: <bucketname>
  s3_bucket_region: eu-central-1
  s3_bucket_subdir: /oracle
  s3_iam_role: <IAM Role for EC2 Instance>

  s3_bucket_subdir is optional. It is needed when the installation media is placed
  in a subdirectory.

Please read the Howto for S3FS before starting with this option!